Rename save_disk to file_operations
