+++ f/common_logic/file_operations.script.yaml
@@ -1,6 +1,6 @@
-summary: 'Shared logic for saving files to disk'
+summary: 'Shared logic for file operations such as extracting files and saving data'
 description: ''
-lock: '!inline f/common_logic/save_disk.script.lock'
+lock: '!inline f/common_logic/file_operations.script.lock'
 has_preprocessor: false
 kind: script
 no_main_func: true
