+++ f/connectors/kobotoolbox/kobotoolbox_responses.py
@@ -10,7 +10,7 @@ from pathlib import Path
 import requests
 
 from f.common_logic.db_operations import StructuredDBWriter, conninfo, postgresql
-from f.common_logic.save_disk import save_data_to_file
+from f.common_logic.file_operations import save_data_to_file
 
 # type names that refer to Windmill Resources
 c_kobotoolbox_account = dict
