+++ tox.ini
@@ -41,6 +41,7 @@ commands =
 [testenv:common_logic]
 deps =
     -r{toxinidir}/f/common_logic/db_transformations.script.lock
+    -r{toxinidir}/f/common_logic/file_operations.script.lock
     -r{toxinidir}/f/common_logic/tests/requirements-test.txt
 commands =
     pytest {posargs} f/common_logic
